test(list): coverage
