--- list/fixtures/custom_element_full.md
@@ -1,58 +1,0 @@
-
-# <my-element>
-
-
-## Attributes
-
-Name        
-my-attr     
-another-attr
-third-attr  
-
-
-## Slots
-
-Name          
-my-slot       
-secondary-slot
-
-
-## Events
-
-Name         | Type       
-my-event     | CustomEvent
-custom-event | CustomEvent
-
-
-## Fields
-
-Name   
-myField
-
-
-## Methods
-
-Name     | Return Type | Privacy | Static
-myMethod | void        | public  | false 
-
-
-## CSS Properties
-
-Name      
-color     
-background
-
-
-## CSS Parts
-
-Name  
-button
-icon  
-
-
-## CSS States
-
-Name    
-open    
-disabled
-
